+++ rosenpass/src/config.rs
@@ -585,204 +585,3 @@ key_out = "/path/to/rp-key-out.txt" # path to store the key
 #peer = "RULdRAtUw7SFfVfGD..." # WireGuard public key
 # extra_params = [] # passed to WireGuard `wg set`
 "###;
-
-#[cfg(test)]
-mod test {
-
-    use super::*;
-    use std::borrow::Borrow;
-
-    fn toml_des<S: Borrow<str>>(s: S) -> Result<toml::Table, toml::de::Error> {
-        toml::from_str(s.borrow())
-    }
-
-    fn toml_ser<S: Serialize>(s: S) -> Result<toml::Table, toml::ser::Error> {
-        toml::Table::try_from(s)
-    }
-
-    fn assert_toml<L: Serialize, R: Borrow<str>>(l: L, r: R, info: &str) -> anyhow::Result<()> {
-        fn lines_prepend(prefix: &str, s: &str) -> anyhow::Result<String> {
-            use std::fmt::Write;
-
-            let mut buf = String::new();
-            for line in s.lines() {
-                writeln!(&mut buf, "{prefix}{line}")?;
-            }
-            Ok(buf)
-        }
-
-        let l = toml_ser(l)?;
-        let r = toml_des(r.borrow())?;
-        ensure!(
-            l == r,
-            "{}{}TOML value mismatch.\n  Have:\n{}\n  Expected:\n{}",
-            info,
-            if info.is_empty() { "" } else { ": " },
-            lines_prepend("    ", &toml::to_string_pretty(&l)?)?,
-            lines_prepend("    ", &toml::to_string_pretty(&r)?)?
-        );
-        Ok(())
-    }
-
-    fn assert_toml_round<'de, L: Serialize + Deserialize<'de>, R: Borrow<str>>(
-        l: L,
-        r: R,
-    ) -> anyhow::Result<()> {
-        let l = toml_ser(l)?;
-        assert_toml(&l, r.borrow(), "Straight deserialization")?;
-
-        let l: L = l.try_into().unwrap();
-        let l = toml_ser(l).unwrap();
-        assert_toml(l, r.borrow(), "Roundtrip deserialization")?;
-
-        Ok(())
-    }
-
-    fn split_str(s: &str) -> Vec<String> {
-        s.split(' ').map(|s| s.to_string()).collect()
-    }
-
-    #[test]
-    fn toml_serialization() -> anyhow::Result<()> {
-        #[cfg(feature = "experiment_api")]
-        assert_toml_round(
-            Rosenpass::empty(),
-            r#"
-            listen = []
-            verbosity = "Quiet"
-            peers = []
-
-            [api]
-            listen_path = []
-            listen_fd = []
-            stream_fd = []
-        "#,
-        )?;
-
-        #[cfg(not(feature = "experiment_api"))]
-        assert_toml_round(
-            Rosenpass::empty(),
-            r#"
-            listen = []
-            verbosity = "Quiet"
-            peers = []
-        "#,
-        )?;
-
-        #[cfg(feature = "experiment_api")]
-        assert_toml_round(
-            Rosenpass::from_sk_pk("/my/sk", "/my/pk"),
-            r#"
-            public_key = "/my/pk"
-            secret_key = "/my/sk"
-            listen = []
-            verbosity = "Quiet"
-            peers = []
-
-            [api]
-            listen_path = []
-            listen_fd = []
-            stream_fd = []
-        "#,
-        )?;
-
-        #[cfg(not(feature = "experiment_api"))]
-        assert_toml_round(
-            Rosenpass::from_sk_pk("/my/sk", "/my/pk"),
-            r#"
-            public_key = "/my/pk"
-            secret_key = "/my/sk"
-            listen = []
-            verbosity = "Quiet"
-            peers = []
-        "#,
-        )?;
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_protocol_version() {
-        let mut rosenpass = Rosenpass::empty();
-        let mut peer_v_02 = RosenpassPeer::default();
-        peer_v_02.protocol_version = ProtocolVersion::V02;
-        rosenpass.peers.push(peer_v_02);
-        let mut peer_v_03 = RosenpassPeer::default();
-        peer_v_03.protocol_version = ProtocolVersion::V03;
-        rosenpass.peers.push(peer_v_03);
-        #[cfg(feature = "experiment_api")]
-        {
-            rosenpass.api.listen_fd = vec![];
-            rosenpass.api.listen_path = vec![];
-            rosenpass.api.stream_fd = vec![];
-        }
-        #[cfg(feature = "experiment_api")]
-        let expected_toml = r#"listen = []
-          verbosity = "Quiet"
-          
-          [api]
-          listen_fd = []
-          listen_path = []
-          stream_fd = []
-
-          [[peers]]
-          protocol_version = "V02"
-          public_key = ""
-
-          [[peers]]
-          protocol_version = "V03"
-          public_key = ""
-          "#;
-        #[cfg(not(feature = "experiment_api"))]
-        let expected_toml = r#"listen = []
-          verbosity = "Quiet"
-
-          [[peers]]
-          protocol_version = "V02"
-          public_key = ""
-
-          [[peers]]
-          protocol_version = "V03"
-          public_key = ""
-          "#;
-        assert_toml_round(rosenpass, expected_toml).unwrap()
-    }
-
-    #[test]
-    fn test_cli_parse_multiple_peers() {
-        let args = split_str(
-            "public-key /my/public-key secret-key /my/secret-key verbose \
-                peer public-key /peer-a/public-key endpoint \
-                peer.test:9999 outfile /peer-a/rp-out \
-                peer public-key /peer-b/public-key outfile /peer-b/rp-out",
-        );
-
-        let config = Rosenpass::parse_args(args).unwrap();
-
-        assert_eq!(
-            config.keypair,
-            Some(RosenpassKeypair::new("/my/public-key", "/my/secret-key"))
-        );
-        assert_eq!(config.verbosity, Verbosity::Verbose);
-        assert!(&config.listen.is_empty());
-        assert_eq!(
-            config.peers,
-            vec![
-                RosenpassPeer {
-                    public_key: PathBuf::from("/peer-a/public-key"),
-                    endpoint: Some("peer.test:9999".into()),
-                    pre_shared_key: None,
-                    key_out: Some(PathBuf::from("/peer-a/rp-out")),
-                    ..Default::default()
-                },
-                RosenpassPeer {
-                    public_key: PathBuf::from("/peer-b/public-key"),
-                    endpoint: None,
-                    pre_shared_key: None,
-                    key_out: Some(PathBuf::from("/peer-b/rp-out")),
-                    ..Default::default()
-                }
-            ]
-        )
-    }
-}
