+++ rosenpass/tests/common/mod.rs
@@ -1,0 +1,2 @@
+mod toml_serialization;
+pub use toml_serialization::assert_toml_round;
